routes and navigation moved out of entry point and renamed so they can access redux store
